+++ CHANGELOG.md
@@ -161,6 +161,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add man pages and command reference to userdocs. #1488
 - Document building images from scratch with Apptainer. #1485
 - Added warewulfd:/overlay-file/{overlay}/{path...}?render={id}
+- Added an alternative image URL /images/{{.ContainerName}} that does not contain a MAC addresses
 
 ### Changed