+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## v4.6.0, unreleased
 
+### Added
+
+- Support Ubuntu-style dracut initrd images.
+
 ### Fixed
 
 - Fix nightly builds.