+++ CHANGELOG.md
@@ -17,6 +17,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fix nightly builds.
+- Better handling for missing NetDev.Device. #1777
+- Remove dracut NetworkManager connections between boot phases. #1780
 
 ### Changed