+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## v4.6.1, unreleased
+
+### Changed
+
+- Update GitHub actions to build aarch64 artifacts.
+
 ## v4.6.0, 2025-03-02
 
 ### Added